+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Next
+### Documentation updates
+- add information about using `match` with input.
+### Examples updates
+- Added `match` example
+
 # 0.5.0
 ### New features
 - Added `.repeat_msg()`.